feat(records): /dashboard/customers records page + sidebar enable
The Customers entry in the sidebar Records group was disabled because no
canonical route existed. Add it.

- `customersRouter.list` paginated procedure (replaces and consolidates
  `leadSourcesRouter.getAllCustomers` — same SQL, no super-admin gate;
  the lead-sources-admin route already gates entry).
- `<CustomersTable>` records component using the canonical pattern
  (`usePaginatedQuery` → `RecordsPageShell` → `QueryToolbar.Standard` →
  `DataTable` with `serverPagination` + `serverSorting`). Mirrors the
  proposals/meetings/projects shape exactly.
- `customer-table-columns.tsx` and `customer-filter-config.ts` moved
  from `features/lead-sources-admin` to `shared/entities/customers/`
  so both the dashboard records page and the lead-sources-admin
  customer panes share one column builder + one filter config. Drops
  the duplicated `CUSTOMER_PAGE_SIZE_OPTIONS` in favor of
  `DEFAULT_RECORDS_PAGE_SIZE_OPTIONS`.
- Sidebar Records → Customers now reads `ability.can('read', 'Customer')`
  instead of being hard-disabled.
