+++ src/shared/entities/customers/components/customer-table-columns.tsx
