+++ src/trpc/routers/customers.router.ts
@@ -1,12 +1,17 @@
 import { TRPCError } from '@trpc/server'
 import { Ratelimit } from '@upstash/ratelimit'
 import { Redis } from '@upstash/redis'
-import { eq, ilike, or } from 'drizzle-orm'
+import { and, desc, eq, gte, ilike, inArray, lte, or } from 'drizzle-orm'
 import z from 'zod'
 import env from '@/shared/config/server-env'
-import { intakeModes } from '@/shared/constants/enums'
+import { customerPipelines, intakeModes } from '@/shared/constants/enums'
 import { getCustomer, getCustomers } from '@/shared/dal/server/customers/api'
 import { addParticipant } from '@/shared/dal/server/meetings/participants'
+import { buildFilterWhere } from '@/shared/dal/server/query/filters'
+import { paginate } from '@/shared/dal/server/query/output'
+import { dateRangeSchema, paginatedQueryInput } from '@/shared/dal/server/query/schemas'
+import { buildSearchWhere } from '@/shared/dal/server/query/search'
+import { buildOrderBy } from '@/shared/dal/server/query/sort'
 import { db } from '@/shared/db'
 import { user } from '@/shared/db/schema/auth'
 import { customerNotes } from '@/shared/db/schema/customer-notes'
@@ -35,6 +40,55 @@ export const customersRouter = createTRPCRouter({
     .query(async ({ ctx }) => {
       const isSuperAdmin = ctx.ability.can('manage', 'all')
       return getCustomers({ isSuperAdmin })
+    }),
+
+  // Server-paginated customers list. Drives /dashboard/customers and the
+  // lead-sources-admin "All customers" pane. Each row carries its joined
+  // leadSource (name + slug); NULL joins mean "unknown legacy import".
+  list: agentProcedure
+    .input(paginatedQueryInput({
+      pipeline: z.array(z.enum(customerPipelines)).optional(),
+      createdAt: dateRangeSchema.optional(),
+    }))
+    .query(async ({ input }) => {
+      const searchWhere = buildSearchWhere(input.search, [customers.name, customers.email])
+      const filterWhere = buildFilterWhere(input.filters, {
+        pipeline: v => (v.length > 0 ? inArray(customers.pipeline, v) : undefined),
+        createdAt: v => and(
+          v.from ? gte(customers.createdAt, v.from) : undefined,
+          v.to ? lte(customers.createdAt, v.to) : undefined,
+        ),
+      })
+      const where = and(searchWhere, filterWhere)
+
+      const orderBy = buildOrderBy(input.sort, {
+        name: customers.name,
+        email: customers.email,
+        createdAt: customers.createdAt,
+        pipeline: customers.pipeline,
+        leadSourceName: leadSourcesTable.name,
+      }, desc(customers.createdAt))
+
+      return paginate({
+        query: () => db
+          .select({
+            id: customers.id,
+            name: customers.name,
+            email: customers.email,
+            createdAt: customers.createdAt,
+            pipeline: customers.pipeline,
+            leadSourceId: customers.leadSourceId,
+            leadSourceName: leadSourcesTable.name,
+            leadSourceSlug: leadSourcesTable.slug,
+          })
+          .from(customers)
+          .leftJoin(leadSourcesTable, eq(leadSourcesTable.id, customers.leadSourceId))
+          .where(where)
+          .orderBy(...orderBy)
+          .limit(input.pagination.limit)
+          .offset(input.pagination.offset),
+        count: () => db.$count(customers, where),
+      })
     }),
 
   // Fetch a single customer by internal UUID
